imooc-llmops-api/requirements-pure.txt
# Web框架核心
flask
flask-login
flask-migrate
flask-cors

# 数据库
sqlalchemy
redis

# AI/ML相关
langchain-core
langchain-openai
langchain-community
tiktoken
faiss-cpu
jieba

# 依赖注入
injector

# 配置管理
python-dotenv

# 模板引擎
jinja2

# HTTP客户端
requests

# 异步任务队列
celery

# 数据验证
pydantic

# JWT认证
PyJWT

# 腾讯云对象存储
cos-python-sdk-v5

# 文件处理
werkzeug

# 测试框架
pytest
